Kk 6 project api

+++ server/controllers/project.ts
@@ -11,18 +11,16 @@ export const createProject = async (request: Request, response: Response) => {
     project.title = request.body.title;
     project.description = request.body.description;
 
-    const userRepository = getRepository(User);
-    const user = await userRepository.findOne(request.session.userId);
-    if (!user) {
-      response.status(401).send();
+    if (response.locals.user) {
+      project.user = response.locals.user;
+      project.assigned_users = [response.locals.user];
+      await projectRepository.save(project);
+      logger.info('Project created successfully: ' + project.id);
+      response.status(201).send({ field: 'alert', message: 'Project successfully created.', project: project });
+    } else {
+      response.status(401).send({ message: 'Access denied' });
       return;
     }
-    project.user = user;
-
-    await projectRepository.save(project);
-    logger.info('Project created successfully: ' + project.id);
-    const _project = { ...project, user: undefined };
-    response.status(201).send({ field: 'alert', message: 'Project successfully created.', project: _project });
   } catch (error) {
     logger.error(error);
     response.status(500).send({ message: error.message });
@@ -31,8 +29,12 @@ export const createProject = async (request: Request, response: Response) => {
 
 export const updateProject = async (request: Request, response: Response) => {
   try {
-    const updatedProject = await response.locals.projectRepository.save({ id: response.locals.project.id, ...request.body });
-    response.status(200).send({ field: 'alert', message: 'Project successfully updated.', project: updatedProject });
+    if (response.locals.projectRepository && response.locals.project) {
+      const updatedProject = await response.locals.projectRepository.save({ id: response.locals.project.id, ...request.body });
+      response.status(200).send({ field: 'alert', message: 'Project successfully updated.', project: updatedProject });
+    } else {
+      response.status(401).send({ field: 'alert', message: 'Access denied' });
+    }
   } catch (error) {
     logger.error(error);
     response.status(500).send({ message: error.message });
@@ -41,8 +43,12 @@ export const updateProject = async (request: Request, response: Response) => {
 
 export const deleteProject = async (_: Request, response: Response) => {
   try {
-    await response.locals.projectRepository.remove(response.locals.project);
-    response.status(200).send({ field: 'alert', message: 'Project successfully deleted.' });
+    if (response.locals.projectRepository && response.locals.project) {
+      await response.locals.projectRepository.remove(response.locals.project);
+      response.status(200).send({ field: 'alert', message: 'Project successfully deleted.' });
+    } else {
+      response.status(401).send({ field: 'alert', message: 'Access denied' });
+    }
   } catch (error) {
     logger.error(error);
     response.status(500).send({ message: error.message });
@@ -51,10 +57,8 @@ export const deleteProject = async (_: Request, response: Response) => {
 
 export const getProjects = async (request: Request, response: Response) => {
   try {
-    const userRepository = getRepository(User);
-    const user = await userRepository.findOne(request.session.userId);
     const projectRepository = getRepository(Project);
-    const projects = await projectRepository.find({ user });
+    const projects: Project[] = await projectRepository.createQueryBuilder('project').leftJoinAndSelect('project.assigned_users', 'user').where(`user.id = ${request.session.userId}`).getMany();
 
     response.status(200).send({ projects });
   } catch (error) {
@@ -64,9 +68,7 @@ export const getProjects = async (request: Request, response: Response) => {
 
 export const getProject = async (request: Request, response: Response) => {
   try {
-    const projectRepository = getRepository(Project);
-    const project = await projectRepository.findOne(request.params.id);
-
+    const project = await response.locals.projectRepository?.findOne(request.params.id, { relations: ['assigned_users'] });
     if (!project) {
       response.status(404).send({ message: 'Project not found' });
       return;
@@ -74,6 +76,61 @@ export const getProject = async (request: Request, response: Response) => {
 
     response.status(200).send({ project });
   } catch (error) {
+    response.status(500).send({ error: error.message });
+  }
+};
+
+export const addUserToProject = async (request: Request, response: Response) => {
+  try {
+    if (!response.locals.project) {
+      response.status(404).send({ message: 'Project not found' });
+      return;
+    }
+
+    const userRepository = getRepository(User);
+    const user = await userRepository.findOne(request.params.uid);
+
+    if (user && response.locals.project) {
+      if (response.locals.project.assigned_users.indexOf(user) > -1) {
+        response.status(403).send({ field: 'alert', message: 'User is already in project' });
+        return;
+      }
+      response.locals.project.assigned_users = [...response.locals.project.assigned_users, user];
+    }
+
+    await response.locals.projectRepository?.save(response.locals.project);
+    logger.info(`User: ${user?.id} added to project: ${response.locals.project.id}`);
+    response.status(200).send({ field: 'alert', message: 'User added to project' });
+  } catch (error) {
+    logger.error(error);
+    response.status(500).send({ error: error.message });
+  }
+};
+
+export const removeUserFromProject = async (request: Request, response: Response) => {
+  try {
+    if (!response.locals.project) {
+      response.status(404).send({ message: 'Project not found' });
+      return;
+    }
+
+    const userRepository = getRepository(User);
+    const user = await userRepository.findOne(request.params.uid);
+
+    if (user) {
+      const userIndex = response.locals.project.assigned_users.indexOf(user);
+      if (userIndex === -1) {
+        response.status(403).send({ field: 'alert', message: 'User is not in project' });
+        return;
+      }
+      response.locals.project.assigned_users = response.locals.project.assigned_users.splice(userIndex, 1);
+    }
+
+    await response.locals.projectRepository?.save(response.locals.project);
+    logger.info(`User: ${user?.id} removed from project: ${response.locals.project.id}`);
+    response.status(200).send({ field: 'alert', message: 'User removed from project' });
+  } catch (error) {
+    logger.error(error);
     response.status(500).send({ error: error.message });
   }
 };
